Add files via upload
This shows virtually every single key factor that has affected the attrition of a company from the Education Field, Age Group, Job Role, Income to the years spent at the company by each individual. It can be deduced that the most attrition was found in the Young Adults category as a result of Low Income that is being paid to them by the company.